Bug fix2 - FfmpegConnector
