+++ app/src/main/java/starklabs/sivodim/Drama/View/HomeActivity.java
@@ -1,13 +1,9 @@
 package starklabs.sivodim.Drama.View;
 
-import android.content.Context;
-import android.content.ContextWrapper;
-import android.content.Intent;
 import android.os.Bundle;
 import android.os.Environment;
 import android.support.design.widget.FloatingActionButton;
 import android.support.design.widget.Snackbar;
-import android.util.Log;
 import android.view.View;
 import android.support.design.widget.NavigationView;
 import android.support.v4.view.GravityCompat;
@@ -18,21 +14,15 @@ import android.support.v7.widget.Toolbar;
 import android.view.Menu;
 import android.view.MenuItem;
 import android.widget.Button;
-import android.widget.ImageView;
-import android.widget.TextView;
 import android.widget.Toast;
 
 import com.github.hiteshsondhi88.libffmpeg.exceptions.FFmpegCommandAlreadyRunningException;
 
 import java.io.File;
-import java.util.Vector;
 
 import starklabs.sivodim.Drama.Model.Screenplay.AudioConcatenator;
 import starklabs.sivodim.Drama.Model.Screenplay.AudioMixer;
-import starklabs.sivodim.Drama.Model.Screenplay.FfmpegStatus;
 import starklabs.sivodim.Drama.Model.Screenplay.Mp3Converter;
-import starklabs.sivodim.Drama.Model.Utilities.Avatar;
-import starklabs.sivodim.Drama.Model.Utilities.Soundtrack;
 import starklabs.sivodim.Drama.Model.Utilities.SpeechSound;
 import starklabs.sivodim.R;
